requirements.txt
appdirs==1.4.4
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.2
defusedxml==0.7.1
distlib==0.3.8
dj-rest-auth==5.0.2
Django==5.0.1
django-allauth==0.60.1
django-cors-headers==4.3.1
django-js-asset==2.2.0
django-mptt==0.16.0
djangorestframework==3.14.0
filelock==3.13.1
idna==3.6
isort==5.13.2
Markdown==3.5.2
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.1.0
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2024.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.2.0
virtualenv==20.25.0
